feat: Move PPO system test to PPO folder.
